Learning about styled-compopnents in SSR components haaaaa
