+++ src/app/page.tsx
@@ -1,9 +1,3 @@
-import { devVars } from './layout'
-
 export default function Home() {
-	return (
-		<main>
-			<h1>{devVars.palette.dark}</h1>
-		</main>
-	)
+	return <main>content goes here</main>
 }
